Create README.md
Scope of The Analysis
To analyze a data set related to FIFA World Cup using a suitable method. In this 
study we have taken up the data sets of the FIFA World Cup(1930-2014) and 
analyzed them using Excel. 
The analysis focused on:
a) which team got the titles of winner,runner-up,third for a desired year
b) Which team won the most titles of winner,runner-up and third
c) country hitting most number of goals per country and number of goals per 
country.
d) Per Year:
• Attendance of audience
• No.of qualified teams
• Goals scored
• Matches played
e) According to attendance:
• Matches with highest of attendance
• Stadium with highest average attendance
 f) Matches outcome according to home and away team